Added Windows script
buddy1.py is the script that runs the robot and the ASR models. It also makes API calls to the TTS container. It's requirements are contained in buddyreqs.txt.
The picoh.py script is from the Picoh program for the picoh robot (pip install picoh) that has been modified by me to work with buddy1.py.